ML_MODEL/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
tensorflow==2.13.0
numpy==1.24.3
pillow==10.4.0
python-multipart==0.0.9

